+++ docs/CHANGELOG.md
@@ -21,7 +21,8 @@
 * Package `dtypes`:
   * Added tests;
   * Added `SizeForDimensions()` to be used for dtypes that uses fractions of bytes (like 4 bits).
-* Added `client.CreateViewOfDeviceBuffer()` to create buffers with shared memory with the host, for faster input.
+* Added `client.NewSharedBuffer` (and the lower level `client.CreateViewOfDeviceBuffer()`) to create buffers with shared 
+  memory with the host, for faster input.
   * Added `AlignedAlloc` and `AlignedFree` required by `client.CreateViewOfDeviceBuffer`.
 
 # v0.4.9 - 2024-11-25